Fixed a defect in MinGW build configuration that crashed the game
The game uses packed structures a lot to save RAM. As it turns out MinGW
ignores the packed attribute on structures by default without the
-mno-ms-bitfields compiler flag. I spent like a month to debug the
issue. It was a nightmare so from now on packed structures shall be
tested with a static_assert for size mismatch.

Also introduced -fno-eliminate-unused-debug-types to instruct the
compiler to keep all type information. There are several type
declarations that are not yet used on C/C++ source code level, but could
be very useful while debugging.