Reimplemented most resource manager functions
The main functionality of the resource manager is to catalog and load
game data from patches.res and max.res on demand.
The original implementation was written in C++ and mixed usage of new,
new[], delete, delete[] with realloc. The original Watcom compiler used
malloc and free as backend for the C++ variants so mixing C++ allocators
with realloc was safe in certain situations. This is not necessarily
true on modern systems. The resource manager does not seem to implement
an actual class or hierarchy of classes. Its just a set of C functions
written in C++ context using new and delete or comparable in addition.
To be able to use realloc and keep implementation close to the original
the module is reimplemented in pure C code and new, delete and such are
replaced by malloc and free.