Reimplemented debug module
The debug module supported the following user interfaces as output:
'mono': monochrome and similar monitors in text mode
'screen': standard output
'log': file log
'gnw': a debug window in-game

The main API function is debug_printf(format, ...) which was probably
abstracted by a macro layer that was disabled for release builds. The
game itself does not use any of the debug services from GNW in v1.00,
v1.04 and v1.56.

The monochrome output interface directly manipulated the display memory
of the hardware which is useless on modern systems so the entire option
was skipped in the reimplementation.