Reimplemented mouse module
The DOS version of GNW originally relied on DOS DPMI services to get the
delta horizontal and vertical mouse movement, as well as button states,
since the last query.
The Windows version of GNW used dxinput_get_mouse_state() which behaved
pretty much the same way.
The mouse hardware abstraction interfaces were replaced with comparable
SDL2 services which is far from ideal at the moment as the event queue
of SDL is now placed in the middle of a getter API replacement. On Linux
this is not that taxing, but on Windows mouse kills performance now.
This could be changed later when the game loop gets reimplemented.

GNW's mouse module supports configuration of mouse sensitivity, but
M.A.X. does not provide a GUI interface to change the setting.

The mouse module supports static and animated cursors, changing cursor
position, setting or getting hot spots and record or play back of mouse
actions through a virtual controller module. Only the left and right
mouse buttons are supported though.