Reimplemented vcr module
The module is able to record and play back human input device events
like keyboard or mouse events. The feature is useful for attract modes,
controlled tutorials or similar, but none of these are used by M.A.X.
Most functions are not even called by the game.

The kb, mouse, vcr modules are integrated and partly abstracted by GNW's
input module. The game itself mostly interacts with the input module
when it comes to control.