Reimplemented window manager module
The window manager (GNW) module is responsible for handling windows and
associated resources like images, textures, buttons and buffers.

The module configuration in M.A.X. supports 50 windows. The manager
originally supported a certain texture format as well, but as M.A.X.
does not use it the feature is currently skipped in the
reimplementation.

Handling of color tables and color effects need to be researched still.
For example the debug module supports a GNW window based operation mode
which turned out to be incompatible with the color management
configuration available in M.A.X. The game does not use the debug module
features actually, but making GNW debug mode work in-game could aid
development in full screen mode.