Reimplemented parts of interface module and did some minor changes
GNW's interface module adds support for dropdown menus, the GNW
windowing based debug window interface, selection list controls and
various timed pop-up message controls.

Most of these are not used by M.A.X. and as the game uses a game
specific color map for color effects and changed some of GNW's related
interfaces further research would be needed to make them work with the
game.

The tm_click_response function is a button on release event handler
which returns a result code in the original implementation. The button
module's ButtonFunc function prototype defines void return type and thus
the button module does not expect any return value. The b_value
parameter is incremented in certain cases within the handler, but the
changed value is lost. As M.A.X. does not use the services that rely on
this handler it is not clear whether this could cause any negative side
effects. The reimplementation skips all return code specific code
snippets that are any ways discarded by the interface.

Many unused complex functions are currently skipped in the
reimplementation.

Updated the code generator to raise GCC diagnostics compatible warnings
when the configuration contains definitions that are not required any
more or when the .in source file still defines already replaced original
function definitions.

Reimplemented the old DOS main function which calls the game main loop.