Disabled VSYNC and enabled HW acceleration for SDL2 renderer
It turned out that SDL2 limits frame rate to the screen refresh rate if
the SDL_RENDERER_PRESENTVSYNC flag is set by SDL_CreateRenderer. M.A.X.
implementes an old fashioned DOS game loop and pushes updates to the
screen as part of core game logic. In other words the renderer and core
game logic are not split into separate threads. Till now SDL2 limited
not only the screen refresh rate, but coincidentally the game logic
cycles per second as well.

On linux game performance was much better so far, but now I think the
reason was that SDL2 on linux probably implements VSYNC in a different
way or not at all.

Finally the game runs smoothly and it was not even necessary to rewrite
the entire DOS game loop for that :)