Reimplemented a couple of game init functions
M.A.X. supported several install sizes. Such a feature is now barely
useful and on case sensitive file systems building paths is not a
trivial matter. SDL2 does not have a CD-ROM abstraction interface
anymore so the related part is just mocked up to a fixed D drive. If
installation type specific command line parameters are not set the game
defaults to full installation in which case the functions do not do
much. Later on the whole complexity could be removed.