Reimplemented assoc module
The associative array module is used by the database manager module in
the DOS version of GNW integrated into M.A.X.

Certain GNW based applications use the associative array module to
manage configuration settings typically stored in human readable form on
the file system. M.A.X. has a dedicated ini file manager for that. Due
to this the module is not really in use. Even GNW's sophisticated
database manager on top of it is replaced by a custom resource manager
which is much simpler.