Reimplemented db module, ditched lzss module
The database manager of GNW is not used by M.A.X. itself. GNW uses it
for various basic file system operations, but not much else.

The database format in this version of GNW is very similar to the one
used by the original Fallout without hashing support. Fallout 2 changed
the database format considerably. The Fallout and M.A.X. database format
relies on the assoc module while Fallout 2 removes that dependency
as-is.

The database module optionally supports lzss compression for textures
and various database resources and would also be able to stream chunks
of decompressed or read uncompressed data from big archives. This was a
very important feature back in the late '90s with the spread of
multimedia in PC gaming. CD-ROMs with their 650 MiB capacity outweighed
typical HDD capcity so games had to access the CD-ROMs a lot as games
simply did not fit onto most hard drives. On the other hand CD-ROM
drives were rather slow. A double speed CD-ROM drive (minimum
requirement for M.A.X.) has a speed rating of 300 KiB/s only. As an
example copying the MASTER.DAT GNW db file of Fallout to the HDD during
installation with a double speed CD-ROM drive would take at least 18
minutes. A M.A.X. review from the time recommends disabling the digital
music tracks, which the reviewer did not like much, as this way the game
runs much smoother from the CD-ROM drive.

As M.A.X. does not have humongous GNW database files and the custom
resource file is 17 MiB only streaming resources, other than digital
music tracks to save RAM space, was not necessary. The reimplementation
skips many unused interfaces and even parts of reimplemented interfaces
for the time being. The unused lzss module is currently removed as-is.