Reimplemented movie module and started work on the mve library modules
The movie module configures the MVE library to play back in-game videos.
The assembly version of the MVE library is modified so that it runs on
modern computers. Currently the MVE library's audio backend is disabled.
The self-modifying image frame decoder algortihms are corrected to
compile with GCC and associated code pages are unlocked for writing by
the self-modifying routines at runtime.
Originally MAX did not support captions for the in-game videos even
though Descent and Fallout realized that feature. The movie module is
implemented in such a way to allow adding captions in the future.

The DOS version of the MVE library had built in audio and video drivers
and the built in video driver supported several rendering methods to
balance performance and quality. All these are removed now and the game
directly renders the frames produced by the MVE library. The audio
interface may be realized with the same approach in the future. MVE's
native video driver also supported scaling the video output which will
be done by SDL in the future; currently there is no upscaling. The
in-game videos have a resolution of 432x320 pixels and the frame rate is
15 FPS. The MVE library supports frame dropping and audio video stream
synchronization but as long as the audio part is missing this feature
had to be stubbed together with the audio.

In short there is still tons of work to do...