Updated ipx module
M.A.X. uses a heavily modified IPX module from Parallax Software
Corporation. Parallax developed Descent and Descent II that were
published by Interplay Productions. Parallax was credited in M.A.X. for
their MS-DOS Installation and Setup Programs, but their network code is
not mentioned anywhere in the game's documentation.

The IPX module relies on Parallax Software's DPMI module to access IPX
networking related DPMI services. The DPMI module built into M.A.X.
defines the following Parallax SW module version ID string: "Id: dpmi.c
1.18 1995/02/02 11:10:22 john Exp". Parallax Software published the
source code of Descent I in 2005 and the DPMI module is newer in it. The
version number and date code are 1.19 and 1995/02/23. This implies that
Interplay took the DPMI driver from Parallax in early 1995 and then did
not care to update it anymore. The 1.19 version fixes a defect in a DPMI
service related functionality. It could be related to DOS real mode
memory management or whatever.

The IPX module of M.A.X. defines the same interfaces that the Parallax
IPX module had. E.g. Descent supported internetwork communication via a
set of APIs and even though M.A.X. only supports a single local network
the internetwork management related APIs are still present in the M.A.X.
executable. It is rather strange that Interplay kept unused APIs, but
they removed the Parallax SW module version ID string... probably with
all the Parallax copyright notices as well.

M.A.X. supports several network interfaces. IPX, modem, serial
communication. The network layer above the IPX module is tightly coupled
with the modem driver. The network interface does not support computer
players. Two human players can play using the modem or serial interfaces
and four players can play via the IPX protocol.

IPX is similar to UDP in the sense that there is no connection
established, broadcasting packets is possible and message sequence and
successful reception are not guaranteed. The IPX module is not error
tolerant at all. The response to most IPX protocol errors is to simply
exit the game. In practice this means that if a player drops from the
game the only way back is that everyone restarts the game and the last
turn's autosave is loaded at match making. Each player's M.A.X. instance
saves the game by itself, there is no such thing like transferring the
save game over the network or similar.

The reimplementation will be based on SDL2_net and UDP. Broadcasting
messages, searching for players and registering them are things to solve
to be able to play over the network. It would be best to keep
compatibility with M.A.X. v1.04 over DOSbox for debugging purposes, but
long term getting rid of all the historical IPX stuff would be best.