Reimplemented ini and inifile modules
The ini module is an INI format parser library written in C that
originally relied on posix file I/O API functions which is odd as its
wrapper uses ISO C file I/O API functions. Probably the module is older
than the rest of the code base and might have been compiled with a
different compiler as it uses cdecl calling convention.

The inifile module is a C++ wrapper for the ini module functions and
related data structure. It implements three classes to manage the three
different type of ini configuration files that the game uses.
MAX.INI: The first ini file type holds all the game settings, 111
entries, and internal parameters that are saved to and loaded from the
file system. The related class still refers to the actual configuration
file as MANDER.INI which implies that it is from the time of
MechaMander. The save game file format also stores the OPTIONS and
PREFERENCES ini sections from this file.
CLANATRB: The second ini file type is a read-only game resource that
holds the clan unit upgrades. The manager class implements an API to
read an attribute called clan gold, which would be added to the starting
gold resource at game setup, but the feature is not used by any of the
clans defined in any of the available M.A.X. versions.
SOUNDVOL: The third ini file type holds sound level or sound volume
information for all digital sound samples of the game. There are 388
entries while the retail version of the M.A.X. CD-ROM contains only 272
digital sound samples. Digital music files are not included.

For the first two types the game creates global objects while the third
one is instantiated on demand. The memory footprint of the third one is
pretty small as it does not create a registry for the sound volume key
value pairs as that is done by the sound manager class.