Integrated the ini and inifile modules
Integrating C++ modules that are not globally instantiated will be
extremely difficult as the size of class objects depend on the compiler
used and either the reimplementations fit into the original space
allocated for them on function stack frames or each affected function
needs to be refactored in the original assembly (by hand). In this
particular case the sound manager instantiates the IniSoundVolumes class
within an initializer function which will be double checked later.