Reimplemented flic manager module
M.A.X. uses FLIC animations for various game assets like unit portraits
or our floating brain in the save/load menu. The format specification
was first published on March 1, 1993 in Dr. Dobb's Journal by Jim Kent
who is the author of Autodesk Animator Pro. The source code of the
software tool is available on github,
https://github.com/AnimatorPro/Animator-Pro . Based on the FLC file
header contents Interplay used this tool to create the FLIC animations.
It is very interesting that many FLC files contain metadata sections
that hint about the authors, their connection to the M.A.X. project and
the source of the art that was exported into FLIC format. For example
the AA_ISO.FLC file contains the string
D:\MECH2\IMAGES\MARKB\AA-FIXED\AA-ISO.PCX which suggests that the art
was created by Mark Brego, one of the 3D animators working on M.A.X.
The module has an interesting hard-coded performance optimization. If
the FLIC animation is loaded from the M.A.X. CD-ROM, then only the first
frame is processed by the game. The animation basically becomes a still
image.