Reimplemented various init and window manager utility functions
The original game verifies the availability or the amount of critical
system resources at startup. E.g. if the drive where the game is
executed from has less than 250 kB free storage space the game warns the
user that saving will be difficult. The limit itself is interesting as I
have save files close to 1 MB in size. The game is not willing to start
up if the free RAM space is less than 6 MB. If physical RAM installed in
the PC is less than 13000 kB, then enhanced graphics is automatically
disabled. If there is no mouse peripheral, the game exits as well. None
of these really matter nowadays... for now.

Finally screenshots are made by M.A.X. instead of GNW matching the
original feature. This changes the format from BMP to PCX.

GNW's windowing system is wrapped by a window manager like interface
similar to the button and image managers. These utility modules
integrate the game assets into GNW. The game menus are realized by C++
classes that all depend on these utility functions. Eventually all of
them are part of the commo.cpp module of the original game. Next step is
to integrate the reimplemented gbutton and gimage modules and to
document the image asset formats that the game stores in RES0 resource
files.

It is very odd that the authors hard-coded the 640x480 screen resolution
in so many places. It is clear that the seamless zoom feature was very
taxing at the time and they did not produce GUI assets for higher
resolutions, but still...