Reimplemented most button manager utility functions
M.A.X. wraps and somewhat misuses GNW's button module. The game specific
module adds texture and sound effect management for GNW buttons and in a
strange way it twists the on click event handler registration scheme.

There are also several cases of potential NULL pointer dereferences and
inappropriate handling of failed dynamic memory allocations. Under DOS
most of those invalid pointer dereferences would not crash the game, but
on modern operating systems they immediately segfault. E.g. the game
will crash if the help button is pushed on the landing zone selection
screen due to such a programming defect.

As of now all on click event handlers are still original code using
Watcom fastcall calling convention so a special adaptation function is
necessary to be able to call them from the reimplemented code. Due to
this interface boundary it will be difficult to reimplement any event
handlers as long as not all of them are reimplemented.