Reimplemented sound manager module
The sound manager module integrated HMI's Sound Operating System (SOS)
v3.x into the game. The SOS supports audio streaming, mixing, panning,
fading, looping between predefined loop points and much more. Digital
music is handled the same way as any other sound sample.

The original M.A.X. mixes maximum 20 audio channels. Most newly
scheduled audio samples are simply skipped if there are no free channels
available at a given time. The more channels there are, the more CPU
resources would be consumed by SOS which was a very big deal back then
even if direct DMA transfers and hand written assembly routines did all
the mixing and related tasks.
All audio samples of the game are in RIFF/WAV format. Most files have a
WAVE, a DATA and a FORMAT chunk only. Some audio files have an optional
SAMPLER chunk that defines a single looping section. The game supports
loading two types of audio files. RIFF/WAV and probably a RAW format
that was not used by the game at the end. This second format parser also
leaks a file handle so it is good that it is not used at the end.
Probably the developers wanted to cut the RIFF/WAV header of the audio
files to keep users from playing them outside the game. E.g. on the
v1.01 retail CD-ROM all non music audio files have the WAV file
extension while on the v1.02 retail CD-ROM the file extension was
changed to SPW.
The sound manager handles three different types of audio samples.
Music files are streamed in small chunks, file management is done by the
game. Even though all music files have a looping section to seamlessly
play a music track forever without the attack phase of the track the
game did not implement streaming the music files and at the same time
handle the loop points. The fact that all music files have loop points
indicates that the original developers planned to add support for
looping music tracks. There is a conventional background music track
shuffler which creates a bit of randomization.
SFX samples all have a predefined volume that is read from an INI
configuration file by the inifile module. SFX samples could also have
loop points and they are used in two ways. SFX samples for ambient sound
of buildings start with a power up sound phase which is cut. Basically
such SFX samples have a start-up phase, a sustain phase and a release
phase and only the sustain phase is kept by the game. All other SFX
samples that have loop points are played from the start and in most
cases the looping section is played as long as a unit is selected by the
mouse. Certain SFX samples are faded out by the game. In every 125 ms
the audio volume is halved until it is muted. The sound manager applies
stereo panning to most SFX audio samples to create a basic positional
audio effect. The further away the screen is from an effect’s source the
weaker the sound volume becomes and if the sound comes from the distant
left then it is played only on the left speaker.
Voice samples of the advisor is the third category. Voices are grouped
into five priority groups. A higher priority voice sample cancels a
lower priority one and if two are scheduled to play soon the higher
priority sample is played first if at all. The five categories in
increasing priority order are 1) on unit select events 2) infiltrator
reaction events 3) response to user actions, commands 4) normal user
alerts to ad hoc events 5) critical user alerts to ad hoc events.

The reimplementation uses SDL_mixer v2.0.4 which does not support loop
points. The latest revision of SDL_mixer adds loop points support to
music tracks only which is not required to match the original game
feature. The mixer does not support streaming of non music mix chunks
out of the box either.
The MVE movie clips have interleaved audio chunks that are streamed on
the fly together with the video stream. As this kind of streaming is not
supported by SDL_mixer there is no audio yet for any of the videos.

These limitations have to be worked around eventually. Or SDL_mixer
needs to be replaced by OpenAL Soft or comparable.