Reimplemented some parts of the base C++ classes
Most C++ objects in M.A.X. are based on smart objects that manage the
lifecycle of the objects and the deallocation of their previously
allocated resources autonomously like a garbage collector would do. In
1995 smart pointers did not exist yet or at least the Watcom 10.5
compiler surely did not implement them so developers had to invent their
own solutions. The compiler also did not support the bool operator and
certain C++ templates specific features. The new and delete operators
only had nothrow variants. As memory allocation failures did not throw,
basically nullptr checks were used in most places. Of course many checks
are missing so if M.A.X. runs out of memory strange things start
happening that no one will be able to understand.

The SmartObject base class and the related SmartPointer class supports
reference counting, various assignment operators, copying and more.
Comparing SmartPointers or checking against nullptr is not as
straightforward in the original design. Maybe it was due to the missing
bool operator. There are two APIs to test for nullptr and both could be
used interchangeably.

The SmartArray is something comparable to a C++ std::vector container.
But the way a SmartArray could be resized is again odd a bit. Typically
the growth factor of a resizable container should be selected in a way
that the deep copy operations of array elements and the rate at which
the capacity of the container needs to be increased are in some kind of
balance. Typically the capacity needs to be doubled, the capacity is
grown exponentially, to achieve this. The SmartArrays instantiated by
the game on the other hand select a fixed growth rate of 5 elements most
of the time. This approach definitively conserves memory at the expense
of the CPU time required to copy around the ever growing container. It
is not the first time that memory efficiency is more important in M.A.X.
than runtime performance.

The various SmartFile objects manage the life cycle of file resources
and allow saving and loading games using serialization and
deserialization of objects. Probably the network code also relies on
this feature. AI tasks, units, complexes (connected buildings creating a
complex), pathfinding lists and many more objects are possible to dump
into binary or text files and to load them up again later. In certain
cases even lazy loading is implemented to reduce game loading times.

There are also smart strings with reference counting and sharing of held
C strings to conserve memory, smart lists, associative arrays or hash
maps or something similar that aim to achieve constant time searching in
these containers and many more.

Reimplementation of all of these smart containers is crucial to be able
to reimplement any higher layer game functionality.