Reimplemented the smart string module
The original module was called strobj. The smart string module
implements reference counting, automatic garbage collection and many
typical string manipulation functions. The container for basic C strings
can be resized if needed. The growth factor is typically 64 characters
(bytes).

The involved classes are much more resource efficient than the standard
String class implementation from the Watcom C/C++ 10.5 compiler.

The smart string class methods are referenced in hundreds of places and
it makes no sense to integrate the reimplementation into the original
code base, but the module is necessary for other smart classes like the
TextFileObject or UNITINFO.

The original SmartString::Substr method is very odd. There is a single
call site in the original code base and it passes the destination
argument by its address without constructing the object first. This
implies that the argument is not a reference, but a pointer to a class
instance. Then in the Substr method itself, the destination is set to
the result of the call to the SmartString class's copy constructor. The
copy constructor was identified via the compiler generated RTTI data.
The class overloads the assignment operator, but that operator is not
called at the place where the copy constructor assigns the object’s
reference to the dereferenced destination argument. The locally copy
constructed object is not destructed as the object goes out of scope
either.

So the caller had a pointer to a not yet allocated object, and the copy
constructor inside the callee is executed without first allocating
memory for the object to be constructed! The placement new operator was
not called either and the object was not destructed when it should have
gone out of scope.

I spent several hours trying to find a way to exactly replicate the
original code, but could not figure it out…

The API also works in a strange way regardless of the above oddities.
The length of the substring to be copied is handled inconsistently
inside the method. In most cases the method copies one more character
then one would expect. E.g. if the API call is string.Substr(position=0,
length=12), then the resulting substring will have 13 characters from
the original string plus the nul C string delimiter.