Reimplemented the sorted array and class register modules
The sorted array is used as an abstract base for various derived
classes. It supports various types of sorting methods via an abstract
SortKey object that derived classes can define. M.A.X. defined two types
of sort keys. One is based on numbers, the other is based on strings.

The MAXRegisterClass and the related RegisterArray classes are used to
hold derived TextFileObject objects. Basically the module implements a
register for various classes that can be loaded from a file or saved to
a file dynamically in binary or human readable text formats. The save
game system uses this feature as it’s basis.