Reimplemented the text file and unitvalues modules
The smart text file reader and writer classes are alternatives to the
binary smart file reader and writer classes. Using the text based
versions allows to produce and process human readable save game and
other file types. The text file classes are the only users of the sorted
enum classes. The retail version of M.A.X. and M.A.X. 2 do not use the
text file classes and as such the sorted enum classes are also stripped
from release builds. Even though these facilities are not currently in
use it is possible to fully restore the functionality.

The unitvalues class manages the unit attributes that are read by the
resource manager class from the ATTRIBS resource file in game. The game
manager class instantiates unitvalues objects for all 93 units.