Reimplemented the ObjectArray and SmartObjectArray containers
The containers are used in various other classes like UnitInfo and
UnitPath to hold structures or basic types. The various instances of
these will further help identification of dependent data structures like
ZoneSquare, Point, MouseEvent, FloodRun and potentially PathSquare or
even more.

There are still some containers missing. One of them is used as some
kind of hash map.

Additionally continued work on TeamUnits and related classes.

Cleaned up the UnitValues::GetAttribute interface. The original game did
not main maintain the textfile based reader and writer for which the
refactoring was necessary. The original game dropped fuel as a unit
attribute in the code while kept it in the ATTIRBS resource file. Added
it back for whatever future shady reasons.