Reimplemented map hash table related classes
The game has two types of hash maps to find UnitInfo class instances
fast. The MapHash uses map grid position as key to find units found at
the given position while the UnitHash uses a unit specific unique id.
The hash function maps keys to the hash table slots and if there are
collisions multiple objects are stored in the table slot in lists. The
table size is configured to 512 items by M.A.X.