Reimplemented various classes for Message Manager
Started to work on the message manager module which depends on many
different GUI specific classes. Most menus are derived from the Window
base class. Windows, dialogs, popup messages, the in-game message and
event logs are all interrelated.

It is very interesting that most menus do not use smart pointers when
they fetch units or other objects and keep those pointers till the life
cycle of the menu object. Maybe this is the root cause for the various
invalid accesses to already deleted objects that crash the game so often
while debug sessions are active under Windows.

Step by step the existing code base above GNW will be converted to
proper C++ to match the original architecture. This is also required to
resolve various interface problems that require lots of dirty hacks
currently to map C++ to C.