Corrected defect fix for null pointer dereference in original game.
This fixes the type selector selection indicator rectangle in unit
purchase and upgrade menus. The proper fix in the C++ reimplementation
will be to wrap the UnitType smart array operator[] call with a smart
array element count test via GetCount() and skip the relevant code if
there are no units to be added to the unit type selector list control.