Started work on network menu.
The modem and null-modem or serial cable networking modes will not be
reimplemented, their respective menus will be scrapped. The IPX
networking mode will be converted to TCP/UDP based without broadcast
messaging. The application layer of the networking protocol, the remote
module, will support a generic transport layer interface which could be
specialized. This abstraction is a tiny bit closer to M.A.X. 2's
architecture, but this does not change the lock-step model nor the
original architecture's weaknesses. The actual transport protocol layer
(reference) implementation will use SDL2_net as backend.