Integration branch for system menus
Watcom C++ is not ABI compatible with GCC. Due to this it is difficult
to integrate and test reimplemented C++ modules in combination with the
original assembly. The game's code is a mass (mess?) of highly coupled
modules that have weak or low cohesion. I hoped to find an interface cut
where the main system menus end and the in-game logic starts, but there
is no easy or meaningful cut. Therefore development will be split into
two directions. The master branch will hold the "working" state of the
game, while this branch will hold the integration of all that is C++
from ground up adding more and more functionality until the two branches
could be merged again.

Currently most system menus are reimplemented and dozens of defect fixes
are present in this branch compared to master. MVE playback is not
working, and all menus stop where the system menu part ends and the
in-game menus and palettes are started to be used.

Linux file system workarounds, and SDL quirks, are not handled within
this branch. It makes no sense to waste time on it now as later any ways
changes will be made to these systems.

Synchronization back to master will depend on how long this integration
branch needs to live.