Reimplemented various classes for upgrade and buy menus.
UnitTypeSelector is the base class for various unit lists like
PurchaseTypeSelector or CargoSelector.

AbstractUpgradeMenu is the base class for menus like the CargoMenu or
the UpgradeMenu. AbstractUpgradeMenu is derived from the Window class
which has an event handler. The Selector could send events to the Menus.

The UnitTypeSelector is the first class that loads in-game resources and
uses the highly optimized assembler sprite scaler routines.