Added support for display selection and aspect ratio correction.
- Preferred monitor display can be selected via max.ini setting
display_index where display_index=0 is the primary monitor.
- Introduced the disable_ar_correction setting in max.ini. The default
value of it is 0 which means that the setting is active. When active the
setting automatically sets the game's aspect ratio to match the selected
display's physical aspect ratio effectively overriding the logical
rederer resolution configured by users by hand. E.g. if user configured
window_width=640 and window_height=480 and the monitor has 16:9 (1.77:1)
aspect ratio than the window_width of 640 will be overridden and set to
853. It is recommended to use screen_mode=2.