Added saved game corruption detection and recovery methods.
- It is potentially possible to recover defect 11, 151 and 183.
- Warning: If corruption is detected at run-time instead of game loading
time before an auto save would be created, the auto save is NOT created.
This behavior is introduced to be able to reload the previous turn in
attempt to be able to inspect potential game events that may lead to the
recurrance of the corruption (that prevented the last auto save).